+++ CHANGELOG.md
@@ -14,10 +14,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
  - #1530, Fix how the PostgREST version is shown in the help text when the `.git` directory is not available - @monacoremo
  - #1094, Fix expired JWTs starting an empty transaction on the db - @steve-chavez
+ - #1475, Fix location header for POST request with select= without PK (#1162) - @wolfgangwalther
 
 ### Changed
 
  - #1522, #1528, #1535, Docker images are now built from scratch based on a the static PostgREST executable (#1494) and with Nix instead of a `Dockerfile`. This reduces the compressed image size from over 30mb to about 4mb - @monacoremo
+ - #1475, Location header for POST request is only included when PK is available on the table (#1461) - @wolfgangwalther
 
 ## [7.0.1] - 2020-05-18